TSP-282 migrate telegramId from Integer to Long (#48)

+++ src/main/java/com/github/sportbot/dto/RegistrationRequest.java
@@ -6,10 +6,10 @@ import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
 
 public record RegistrationRequest(
-    @NotNull Integer telegramId,
-    @NotNull String sendPulseId,
-    boolean isSubscribed,
-    @NotBlank String fullName,
-    Integer referrerTelegramId,
-    LocalTime remindTime
+        @NotNull Long telegramId,
+        @NotNull String sendPulseId,
+        boolean isSubscribed,
+        @NotBlank String fullName,
+        Integer referrerTelegramId,
+        LocalTime remindTime
 ) {}
